requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backoff==2.2.1
bcrypt==4.0.1
boto3==1.28.76
botocore==1.31.76
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
coloredlogs==15.0.1
deepgram-sdk==2.11.0
Delorean==1.0.0
Deprecated==1.2.14
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gensim==4.3.2
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
hubspot==0.1.14.dev0
huggingface-hub==0.17.3
humanfriendly==10.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jmespath==1.0.1
joblib==1.3.2
kubernetes==28.1.0
libretranslatepy==2.1.1
lxml==4.9.3
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.28.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.2
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.14.1
tqdm==4.66.1
translate==3.6.1
typer==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
